+++ src/nightwatch/nightwatch.config.js
@@ -11,7 +11,7 @@ const seleniumHost = ip.address();
 
 const nightwatchConfig = (webpackConfig, srcFolders, providedPort) => {
   // eslint-disable-next-line no-console
-  console.error('WARNING: WebdriverIO should be used instead of Nightwatch.  Nightwatch will be deprecated in the future.');
+  console.warn('WARNING: WebdriverIO should be used instead of Nightwatch.  Nightwatch will be deprecated in the future.');
 
   if (providedPort) {
     port = providedPort;