+++ wdio.conf.js
@@ -1,19 +1,18 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 const wdioConf = require('./lib/wdio/conf');
-const WebpackDevService = require('./lib/wdio/services/index').WebpackDevService;
+const ExpressDevService = require('./lib/wdio/services/index').ExpressDevService;
 const webpackConfig = require('./tests/test.config.js');
 const localIP = require('ip');
 const path = require('path');
 
-const webpackPort = 8080;
+const port = 8080;
 
 const config = {
   ...wdioConf.config,
-  baseUrl: `http://${localIP.address()}:${webpackPort}`,
+  baseUrl: `http://${localIP.address()}:${port}`,
   specs: [path.join('.', 'tests', 'wdio', '**-spec.js')],
 
-  // Configuration for WebpackDevService
-  webpackPort,
+  // Configuration for ExpressDevService
   webpackConfig,
 
   // Configuration for SeleniumDocker service
@@ -32,6 +31,5 @@ const config = {
   },
 };
 
-
-config.services = wdioConf.config.services.concat([WebpackDevService]);
+config.services = wdioConf.config.services.concat([ExpressDevService]);
 exports.config = config;